+++ docs/griptape-framework/misc/src/events_3.py
@@ -1,7 +1,5 @@
-from typing import cast
-
 from griptape.drivers import OpenAiChatPromptDriver
-from griptape.events import CompletionChunkEvent, EventBus, EventListener
+from griptape.events import EventBus, EventListener, TextChunkEvent
 from griptape.structures import Pipeline
 from griptape.tasks import ToolkitTask
 from griptape.tools import PromptSummaryTool, WebScraperTool
@@ -9,8 +7,8 @@ from griptape.tools import PromptSummaryTool, WebScraperTool
 EventBus.add_event_listeners(
     [
         EventListener(
-            lambda e: print(cast(CompletionChunkEvent, e).token, end="", flush=True),
-            event_types=[CompletionChunkEvent],
+            lambda e: print(e.token, end="", flush=True),
+            event_types=[TextChunkEvent],
         )
     ]
 )