+++ griptape/events/base_chunk_event.py
@@ -1,0 +1,8 @@
+from attrs import define, field
+
+from griptape.events.base_event import BaseEvent
+
+
+@define
+class BaseChunkEvent(BaseEvent):
+    index: int = field(default=0, metadata={"serializable": True})